# model/entity/employe.go
package entity

import (
	"time"

	"gorm.io/gorm"
)

type Employe struct {
	ID        uint           `json:"id" gorm:"primaryKey"`
	Name      string         `json:"name"`
	Position  string         `json:"position"`
	Salary    int            `json:"salary"`
	CreatedAt time.Time      `json:"created_at"`
	UpdatedAt time.Time      `json:"updated_at"`
	DeletedAt gorm.DeletedAt `json:"-" gorm:"index,column:deleted_at"`
}